# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "video-intelligence-platform",
                 "description":  "Local Video Intelligence Platform - GPU-constrained video analysis",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/queries":  {
                                   "post":  {
                                                "tags":  [
                                                             "queries"
                                                         ],
                                                "summary":  "Query Videos",
                                                "description":  "Query video content using natural language.",
                                                "operationId":  "query_videos_queries_post",
                                                "requestBody":  {
                                                                    "content":  {
                                                                                    "application/json":  {
                                                                                                             "schema":  {
                                                                                                                            "$ref":  "#/components/schemas/QueryRequest"
                                                                                                                        }
                                                                                                         }
                                                                                },
                                                                    "required":  true
                                                                },
                                                "responses":  {
                                                                  "200":  {
                                                                              "description":  "Successful Response",
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/QueryResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                          }
                                                                          },
                                                                  "422":  {
                                                                              "description":  "Validation Error",
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                  }
                                                                                                                   }
                                                                                          }
                                                                          }
                                                              }
                                            }
                               },
                  "/queries/search":  {
                                          "get":  {
                                                      "tags":  [
                                                                   "queries"
                                                               ],
                                                      "summary":  "Search Transcripts",
                                                      "description":  "Search video transcripts directly.",
                                                      "operationId":  "search_transcripts_queries_search_get",
                                                      "parameters":  [
                                                                         {
                                                                             "name":  "q",
                                                                             "in":  "query",
                                                                             "required":  true,
                                                                             "schema":  {
                                                                                            "type":  "string",
                                                                                            "title":  "Q"
                                                                                        }
                                                                         },
                                                                         {
                                                                             "name":  "video_id",
                                                                             "in":  "query",
                                                                             "required":  false,
                                                                             "schema":  {
                                                                                            "anyOf":  [
                                                                                                          {
                                                                                                              "type":  "string",
                                                                                                              "format":  "uuid"
                                                                                                          },
                                                                                                          {
                                                                                                              "type":  "null"
                                                                                                          }
                                                                                                      ],
                                                                                            "title":  "Video Id"
                                                                                        }
                                                                         },
                                                                         {
                                                                             "name":  "limit",
                                                                             "in":  "query",
                                                                             "required":  false,
                                                                             "schema":  {
                                                                                            "type":  "integer",
                                                                                            "default":  10,
                                                                                            "title":  "Limit"
                                                                                        }
                                                                         }
                                                                     ],
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "Successful Response",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "type":  "object",
                                                                                                                                            "additionalProperties":  true,
                                                                                                                                            "title":  "Response Search Transcripts Queries Search Get"
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                },
                                                                        "422":  {
                                                                                    "description":  "Validation Error",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                }
                                                                    }
                                                  }
                                      },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "QueryRequest":  {
                                                            "properties":  {
                                                                               "query":  {
                                                                                             "type":  "string",
                                                                                             "maxLength":  1000,
                                                                                             "minLength":  1,
                                                                                             "title":  "Query"
                                                                                         },
                                                                               "video_id":  {
                                                                                                "anyOf":  [
                                                                                                              {
                                                                                                                  "type":  "string",
                                                                                                                  "format":  "uuid"
                                                                                                              },
                                                                                                              {
                                                                                                                  "type":  "null"
                                                                                                              }
                                                                                                          ],
                                                                                                "title":  "Video Id"
                                                                                            }
                                                                           },
                                                            "type":  "object",
                                                            "required":  [
                                                                             "query"
                                                                         ],
                                                            "title":  "QueryRequest"
                                                        },
                                       "QueryResponse":  {
                                                             "properties":  {
                                                                                "answer":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Answer"
                                                                                           },
                                                                                "sources":  {
                                                                                                "items":  {
                                                                                                              "$ref":  "#/components/schemas/QuerySource"
                                                                                                          },
                                                                                                "type":  "array",
                                                                                                "title":  "Sources"
                                                                                            },
                                                                                "confidence":  {
                                                                                                   "type":  "number",
                                                                                                   "title":  "Confidence"
                                                                                               },
                                                                                "cached":  {
                                                                                               "type":  "boolean",
                                                                                               "title":  "Cached",
                                                                                               "default":  false
                                                                                           }
                                                                            },
                                                             "type":  "object",
                                                             "required":  [
                                                                              "answer",
                                                                              "sources",
                                                                              "confidence"
                                                                          ],
                                                             "title":  "QueryResponse"
                                                         },
                                       "QuerySource":  {
                                                           "properties":  {
                                                                              "video_id":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Video Id"
                                                                                           },
                                                                              "text":  {
                                                                                           "type":  "string",
                                                                                           "title":  "Text"
                                                                                       },
                                                                              "start_ms":  {
                                                                                               "type":  "integer",
                                                                                               "title":  "Start Ms"
                                                                                           },
                                                                              "end_ms":  {
                                                                                             "type":  "integer",
                                                                                             "title":  "End Ms"
                                                                                         },
                                                                              "score":  {
                                                                                            "type":  "number",
                                                                                            "title":  "Score"
                                                                                        }
                                                                          },
                                                           "type":  "object",
                                                           "required":  [
                                                                            "video_id",
                                                                            "text",
                                                                            "start_ms",
                                                                            "end_ms",
                                                                            "score"
                                                                        ],
                                                           "title":  "QuerySource"
                                                       },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           },
...
                                   }
                   }
}